+++ cdot65/prisma_access/plugins/module_utils/api_spec.py
@@ -140,7 +140,7 @@ class PrismaAccessSpec:
         )
 
     @staticmethod
-    def address_groups_spec():
+    def address_group_spec():
         """Return the address groups object spec."""
         return dict(
             description=dict(